group fixtures for accesing tenant wallets. renamed for clarity
